WaveSyn console now has a steady auto completion engine.
Originally, I tried to utilize the code from idlelib to implement auto
completion, which is not recommanded because idlelib is exclusively for
idle. Today I found the module Jedi which is an easy to use auto
completion engine, and now I use it to implement the auto completion of
WaveSyn console.